feat: customer address now uses gate to authorize

+++ app/Http/Controllers/customer/CustomerAddressController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers\customer;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreUserAddressRequest;
 use App\Http\Resources\CustomerAddressResource;
+use App\Models\Customer;
 use App\Models\CustomerAddress;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 
 class CustomerAddressController extends Controller
@@ -13,8 +15,8 @@ class CustomerAddressController extends Controller
     public function index()
     {
         $user = auth()->user()->id;
-        return 
-        CustomerAddressResource::collection(CustomerAddress::where('user_id', $user)->get());
+        return
+            CustomerAddressResource::collection(CustomerAddress::where('user_id', $user)->get());
     }
 
     public function store(StoreUserAddressRequest $request)
@@ -34,36 +36,62 @@ class CustomerAddressController extends Controller
 
     public function show($id)
     {
-        $user = auth()->user()->id;
-        $address = CustomerAddress::where('user_id', $user)
-            ->find($id)
-            ->get();
+        $customerAddress = CustomerAddress::find($id);
+        if (!Gate::allows('customerAddress', $customerAddress)) {
+            return response('آیدی وارد شده آیدی آدرس های شما نیست', 401);
+        }
 
-        return CustomerAddressResource::collection($address);
+        $customerAddress = CustomerAddress::where('id', $id)->get();
+        return CustomerAddressResource::collection($customerAddress);
     }
 
     public function update(StoreUserAddressRequest $request, $id)
     {
         $request->validated();
 
-        $user_id = auth()->user()->id;
-        $address = CustomerAddress::where('user_id', $user_id)
-            ->get()
-            ->find($id);
+        $customerAddress = CustomerAddress::find($id);
+        if (!Gate::allows('customerAddress', $customerAddress)) {
+            return response('آیدی وارد شده آیدی آدرس های شما نیست', 401);
+        }
 
-        $address->update([
+        $customerAddress->update([
             'title' => $request->input('title'),
             'address' => $request->input('address'),
             'latitude' => $request->input('latitude'),
             'longitude' => $request->input('longitude'),
         ]);
 
-        return 'با موفقیت تغییر داده شد';
+        return response('با موفقیت تغییر داده شد', 200);
     }
 
     public function destroy($id)
     {
+        $customerAddress = CustomerAddress::find($id);
+        if (!Gate::allows('customerAddress', $customerAddress)) {
+            return response('آیدی وارد شده آیدی آدرس های شما نیست', 401);
+        }
+        
+        //200 or 204?
         CustomerAddress::destroy($id);
-        return 'با موفقیت پاک شد';
+        return response('با موفقیت پاک شد', 200);
+    }
+
+    public function currentAddress($id)
+    {
+
+        $user_id = auth()->user()->id;
+
+        $customerAddress = CustomerAddress::find($id);
+        if (!Gate::allows('customerAddress', $customerAddress)) {
+            return response('آیدی وارد شده آیدی آدرس های شما نیست', 401);
+        }
+
+        $currentAddress = Customer::find($user_id)
+            ->update([
+                "currentAddress" => $customerAddress
+            ]);
+
+        return response('آدرس فعلی شما با موفقیت ثبت شد',200);
+        // return CustomerAddressResource::collection(Customer::find($user_id)->currentAddress);
     }
 }
